requirements.txt
rich
pyYAML
